Change pubtime and eventtime to be long long.

+++ include/lsst/ctrl/events/Event.h
@@ -64,8 +64,8 @@ public:
     PropertySet::Ptr getPropertySet() const;
 
     std::string getPubDate();
-    long getPubTime();
-    void setPubTime(long t);
+    long long getPubTime();
+    void setPubTime(long long t);
 
     long getEventTime();
     std::string getEventDate();
@@ -92,13 +92,13 @@ protected:
     void _constructor(const std::string& runid, const PropertySet& ps);
     virtual void setKeywords(PropertySet::Ptr psp) const;
 
-    long _eventTime;
+    long long _eventTime;
     std::string _type;
     std::string _topic;
     std::string _hostId;
     std::string _runId;
     std::string _status;
-    long _pubTime;
+    long long _pubTime;
 /*
 protected:
     void init(const std::string& runId);
